+++ sklearn/metrics/_base.py
@@ -19,6 +19,7 @@ import numpy as np
 from ..base import is_classifier
 from ..utils import check_array, check_consistent_length
 from ..utils.multiclass import type_of_target
+from ..utils.validation import _check_classifier_response_method
 
 
 def _average_binary_score(binary_metric, y_true, y_score, average,
@@ -252,58 +253,6 @@ def _check_pos_label_consistency(pos_label, y_true):
     return pos_label
 
 
-def _check_classifier_response_method(estimator, response_method):
-    """Return prediction method from the `response_method`.
-
-    Parameters
-    ----------
-    estimator : estimator instance
-        Classifier to check.
-
-    response_method : {'auto', 'predict_proba', 'decision_function', 'predict'}
-        Specifies whether to use :term:`predict_proba` or
-        :term:`decision_function` as the target response. If set to 'auto',
-        :term:`predict_proba` is tried first and if it does not exist
-        :term:`decision_function` is tried next and :term:`predict` last.
-
-    Returns
-    -------
-    prediction_method : callable
-        Prediction method of estimator.
-    """
-
-    possible_response_methods = (
-        "predict", "predict_proba", "decision_function", "auto"
-    )
-    if response_method not in possible_response_methods:
-        raise ValueError(
-            f"response_method must be one of "
-            f"{','.join(possible_response_methods)}."
-        )
-
-    error_msg = "response method {} is not defined in {}"
-    if response_method != "auto":
-        prediction_method = getattr(estimator, response_method, None)
-        if prediction_method is None:
-            raise ValueError(
-                error_msg.format(response_method, estimator.__class__.__name__)
-            )
-    else:
-        predict_proba = getattr(estimator, 'predict_proba', None)
-        decision_function = getattr(estimator, 'decision_function', None)
-        predict = getattr(estimator, 'predict', None)
-        prediction_method = predict_proba or decision_function or predict
-        if prediction_method is None:
-            raise ValueError(
-                error_msg.format(
-                    "decision_function, predict_proba or predict",
-                    estimator.__class__.__name__
-                )
-            )
-
-    return prediction_method
-
-
 def _get_response(
     estimator,
     X,
@@ -348,11 +297,11 @@ def _get_response(
     """
     if is_classifier(estimator):
         y_type = type_of_target(y_true)
-        classes = estimator.classes_
         prediction_method = _check_classifier_response_method(
             estimator, response_method
         )
         y_pred = prediction_method(X)
+        classes = estimator.classes_
 
         if pos_label is not None and pos_label not in classes:
             raise ValueError(
