+++ sklearn/metrics/_plot/roc_curve.py
@@ -210,7 +210,8 @@ def plot_roc_curve(estimator, X, y, *, sample_weight=None,
     check_matplotlib_support('plot_roc_curve')
 
     y_pred, pos_label = _get_response(
-        X, estimator, response_method, pos_label=pos_label)
+        estimator, X, y, response_method, pos_label=pos_label
+    )
 
     fpr, tpr, _ = roc_curve(y, y_pred, pos_label=pos_label,
                             sample_weight=sample_weight,
